Renaming spark_delta => spark
