+++ src/cdm_data_loaders/parsers/uniprot/idmapping.py
@@ -37,7 +37,7 @@ from cdm_data_loaders.core.constants import CDM_LAKE_S3, INVALID_DATA_FIELD_NAME
 from cdm_data_loaders.core.pipeline_run import PipelineRun
 from cdm_data_loaders.readers.dsv import read
 from cdm_data_loaders.utils.s3 import list_matching_objects
-from cdm_data_loaders.utils.spark_delta import APPEND, set_up_workspace, write_table
+from cdm_data_loaders.utils.spark import APPEND, set_up_workspace, write_table
 from cdm_data_loaders.validation.dataframe_validator import DataFrameValidator, Validator
 from cdm_data_loaders.validation.df_nullable_fields import validate as check_nullable_fields
